# tarot/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample API for learning microservices",
        "title": "Microsvc-dd",
        "contact": {},
        "version": "0.1.0-rc"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/tarots/{cardnumber}": {
            "get": {
                "description": "Get a tarot card by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tarot"
                ],
                "summary": "Get a tarot card by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tarot ID",
                        "name": "cardnumber",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TarotModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.TarotModel": {
            "description": "Tarot Model",
            "type": "object",
            "required": [
                "cardname",
                "number",
                "type"
            ],
            "properties": {
                "cardname": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "The Fool"
                },
                "number": {
                    "type": "string",
                    "maxLength": 15,
                    "example": "0"
                },
                "type": {
                    "type": "string",
                    "maxLength": 63,
                    "example": "major"
                }
            }
        }
    }
}